Various FS work, Work on some demo_level stuff
Added check on player to update surface type with a line trace. The update surface type funciton can be called in the IK update once it is implemented.

Added an audio volume to primary building so I can modify ambiance based on when you're inside/outside. Not functional yet, needs to be finished.

Added FS notify for player landing

Added concrete footstep sounds